feat: database structure refactor
